### 2021-01-06 #2942 @vavrines
> @whedon query scope
>
> @vavrines thanks for your submission to JOSS – due to the relatively small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week. If you want to fix the DOIs (noting that whedon's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
>
> @whedon generate pdf
> @whedon check references

Thanks for the prompt support!
Maybe I should mention it here. The package is a combination that reexports [KitBase](https://github.com/vavrines/KitBase.jl), [KitML](https://github.com/vavrines/KitML.jl) and [KitFort](https://github.com/vavrines/KitFort.jl), and the total size of codes should be fine : )
