### 2021-01-06 #2942 @vavrines
> > Maybe I should mention it here. The package is a combination that reexports KitBase, KitML and KitFort, and the total size of codes should be fine : )
>
> @vavrines - I'm not sure I follow sorry. Are you saying that this package is actually four different codebases combined, or that KitBase, KitML, and KitFort are _dependencies_ of this one?
>
> If so, JOSS' scholarly effort criterion is applied to the package under review (Kinetic.jl) in this case, not the dependent projects. From my assessment (with the caveat that I'm not a Julia user), this submission looks to be ~800 lines of Julia code. Please correct me if I'm missing something here.

@arfon Thanks for the clarification.
Yes they're the codebases. The structs and methods have been imported into this main package so there's no need to review these sub repositories.
It's a common strategy in Julia to reduce precompilation time, e.g. [DIfferentialEquations.jl](https://github.com/SciML/DifferentialEquations.jl) and [Yao.jl](https://github.com/QuantumBFS/Yao.jl)
