+++ docs/src/index.md
@@ -1,22 +1,28 @@
 # Kinetic.jl
 
-Kinetic.jl is a lightweight toolbox for modeling and simulation on the basis of kinetic theory of gases, photons, plasmas and neutrons.
-It can be used to solve either Boltzmann and its related model equations
+Kinetic is a lightweight Julia toolbox for the study of computational fluid dynamics and scientific machine learning.
+The main module consists of [KitBase.jl](https://github.com/vavrines/KitBase.jl) with basic physics and [KitML.jl](https://github.com/vavrines/KitML.jl) with neural dynamics. 
+The high-performance Fortran library [KitFort.jl](https://github.com/vavrines/KitFort.jl) is optional and can be manually imported when the executing efficiency becomes the first priority. 
+A wrapper [kineticpy](https://github.com/vavrines/kineticpy) has been built to locate the data hierarchies and methods in Python.
+
+Kinetic is interested in theoretical modeling and numerical simulation of many-particle systems, e.g. gases, photons, plasmas, neutrons, electrons, etc.
+It employs the finite volume method (FVM) to conduct 1-3 dimensional numerical simulations on CPUs and GPUs.
+Any advection-diffusion type equation can be hooked and solved within the framework.
+The user interface is designed as intuitive and neat as possible, allowing users to focus on physics and to cooperate with existing packages in the Julia ecosystem.
+We pay special attentions to Hilbert's sixth problem, i.e. building the numerical passage between the kinetic theory of gases and continuum mechanics. 
+Two types of governing equations are therefore bridged, i.e. the Boltzmann equation
 
 ```math
 \frac{\partial f}{\partial t}+ \mathbf u \cdot \nabla_\mathbf x f + \mathbf a \cdot \nabla_\mathbf u f = \int_{\mathcal R^3} \int_{\mathcal S^2} B (f(\mathbf u_*')f(\mathbf u')-f(\mathbf u_*) f(\mathbf u)) d\Omega d\mathbf u_*
 ```
 
-or their upscaling moment system
+and its upscaling moment system
 
 ```math
 \frac{\partial \mathbf W}{\partial t} + \nabla_\mathbf x \cdot \mathbf F = \mathbf S
 ```
 
-The solution algorithms are organized using the finite volume method (FVM), and can be run in 1-3 dimensions on CPUs and GPUs.
-The user interface is designed as intuitive and neat as possible, allowing users to focus on physics and to cooperate with existing packages in the Julialang ecosystem.
-
-# Getting help
+## Getting help
 
 If you are interested in using Kinetic.jl or are trying to figure out how to use it, please feel free to get in touch and raise questions.
 Open an issue or pull request if you have questions, suggestions or solutions.
