### 2021-01-15 0fa3569
Add Basics in docs
