+++ docs/src/basics.md
@@ -1,33 +1,75 @@
 # Basics
 
-The real world is built upon physics, which show a diverse set of behaviors on different characteristic scales.
+The physical world shows a diverse set of behaviors on different characteristic scales.
 Consider the molecular motion of gases as an example.
-Down to finest scale of characteristic motion, the Newton's second law depicts particle motions
+Down to the finest scale of a many-particle system, the Newton's second law depicts particle motions via
 ```math
 \mathbf{F} = m \mathbf{a}
 ```
-As a first order system this reads
+
+As a first order system it reads
 ```math
-\frac{d \mathbf x}{dt} = \mathbf v, \frac{d \mathbf v}{dt} = \frac{\mathbf F}{m}
+\frac{d \mathbf x}{dt} = \mathbf v, \ \frac{d \mathbf v}{dt} = \frac{\mathbf F}{m}
 ```
 
-There are more than `2e25` molecules per cubic meter in normal atmosphere.
-As a result, it is despairing to track all the individual motions every moment.
-Instead, ensemble averaging is needed to provide a bottom-up view, from which the coarse-grained modeling happens.
+An intuitive numerical algorithm is to get the numerous particles on board and track the trajectories of them.
+A typical example is the [Molecular Dynamics](https://en.wikipedia.org/wiki/Molecular_dynamics).
+This is not going to be efficient since there are more than `2e25` molecules per cubic meter in normal atmosphere, and things get even more complicated when you count on the N-body interactions all the time.
+Some methods have been proposed to simplify the computation.
+As an example, the [Direct simulation Monte Carlo](https://en.wikipedia.org/wiki/Direct_simulation_Monte_Carlo) employs certain molecular models and conduct the intermolecular collisions in a stochastic manner.
+It significantly reduces the computational cost, while the trade-off is the artificial fluctuations.
+Many realizations must be simulated successively to average the solutions and reduce the errors.
+
+
+An alternative strategy is made from ensemble averaging, where the coarse-grained modeling is used to provide a bottom-up view.
 At the mean free path and collision time scale of molecules, particles travel freely during most of time with mild intermolecular collisions.
-Such dynamics can be described with an operator splitting approach, i.e. the Boltzmann equation
-
+Such dynamics can be described with an operator splitting approach, i.e. the kinetic transport equation
 ```math
-\frac{\partial f}{\partial t}+ \mathbf u \cdot \nabla_\mathbf x f + \mathbf a \cdot \nabla_\mathbf u f = \int_{\mathcal R^3} \int_{\mathcal S^2} B (f(\mathbf u_*')f(\mathbf u')-f(\mathbf u_*) f(\mathbf u)) d\Omega d\mathbf u_*
+\frac{\partial f}{\partial t}+ \mathbf v \cdot \nabla_\mathbf x f + \mathbf a \cdot \nabla_\mathbf v f = Q(f)
 ```
+where the left and right hand sides model particle transports and collisions correspondingly.
+Different collision models can be inserted into such equation.
+If the particles only collide with a background material one obtains linear Boltzmann collision operator
+```math
+Q(f)=\int_{\mathbb R^3} \mathcal B(\mathbf v_*, \mathbf v) \left[ f(\mathbf v_*)-f(\mathbf v)\right] d\mathbf v_*
+```
+where the collision kernel `B` models the strength of collisions at different velocities. 
+If the interactions among particles are considered, the collision operator becomes nonlinear. 
+For example, the two-body collision results in nonlinear Boltzmann equation
+```math
[…]
