+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api_theory/#Theory","page":"Theory","title":"Theory","text":"","category":"section"},{"location":"api_theory/","page":"Theory","title":"Theory","text":"prim_conserve\nconserve_prim\nmixture_prim_conserve\nmixture_conserve_prim\nem_coefficients\nadvection_flux\nburgers_flux\neuler_flux\neuler_jacobi\ngauss_moments\nmixture_gauss_moments\nmoments_conserve\nmixture_moments_conserve\npdf_slope\nmixture_pdf_slope\nmoments_conserve_slope\nmixture_moments_conserve_slope\ndiscrete_moments\nstress\nheat_flux\nmaxwellian\nmixture_maxwellian\nshakhov\nreduce_distribution\nfull_distribution\nref_vhs_vis\nvhs_collision_time\naap_hs_collision_time\naap_hs_prim\naap_hs_diffeq!\nshift_pdf!\nhs_boltz_kn\nkernel_mode\nboltzmann_fft\nboltzmann_fft!\nheat_capacity_ratio\nsound_speed","category":"page"},{"location":"api_theory/#KitBase.prim_conserve","page":"Theory","title":"KitBase.prim_conserve","text":"Transform primitive -> conservative variables\n\nprim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,1}}\n\nprim_conserve(ρ, U, λ, γ)\n\nprim_conserve(ρ, U, V, λ, γ)\n\nprim_conserve(ρ, U, V, W, λ, γ)\n\n\n\n\n\n","category":"function"},{"location":"api_theory/#KitBase.conserve_prim","page":"Theory","title":"KitBase.conserve_prim","text":"Transform conservative -> primitive variables\n\nscalar: pseudo primitive vector for scalar conservation laws\nconserve_prim(u)\nconserve_prim(u, a)\nvector: primitive vector for Euler, Navier-Stokes and extended equations\nconserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,1}}\nconserve_prim(ρ, M, E, γ)\nconserve_prim(ρ, MX, MY, E, γ)\n\n\n\n\n\n","category":"function"},{"location":"api_theory/#KitBase.mixture_prim_conserve","page":"Theory","title":"KitBase.mixture_prim_conserve","text":"Transform multi-component primitive -> conservative variables\n\nmixture_prim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,2}}\n\n\n\n\n\n","category":"function"},{"location":"api_theory/#KitBase.mixture_conserve_prim","page":"Theory","title":"KitBase.mixture_conserve_prim","text":"Transform multi-component conservative -> primitive variables\n\nmixture_conserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,2}}\n\n\n\n\n\n","category":"function"},{"location":"api_theory/#KitBase.em_coefficients","page":"Theory","title":"KitBase.em_coefficients","text":"Calculate electromagnetic coeffcients in hyperbolic Maxwell's equations\n\nem_coefficients(\n    prim::X,\n    E::Y,\n    B::Z,\n    mr,\n    lD,\n    rL,\n    dt,\n) where {X<:AbstractArray{<:Real,2},Y<:AbstractArray{<:Real,1},Z<:AbstractArray{<:Real,1}}\n\n\n\n\n\n","category":"function"},{"location":"api_theory/#KitBase.advection_flux","page":"Theory","title":"KitBase.advection_flux","text":"Theoretical flux of linear advection equation\n\nadvection_flux(u, a)\n\n\n\n\n\n","categ
[…]
