+++ docs/src/solver.md
@@ -1,12 +1,28 @@
 # Solution Algorithm
 
-The finite volume method (FVM) is employed in this package. The general solution algorithm can be conclude as follows, where both explicit and implicit methods are implemented.
+The finite volume method (FVM) is employed in Kinetic. 
+The general solution algorithm can be conclude as follows, where both explicit and implicit methods are implemented.
 
 ![](./assets/solver_process.png)
 
+The high-level solver function is 
+```@docs
+KitBase.solve!
+```
+
+The detailed solution procedures can be concluded as follows
+- pre-process
+- timestep calculation
+- reconstruction
+- evolution
+- update
+- post-process
+
+
+
 (1) pre-process: `initialize(configfilename::AbstractString)`
 * new run: .txt / .cfg / .toml
-* restart: .jld2 (HDF5) -> solSet, ctr, face, t
+* restart: .jld2 (HDF5) -> SolverSet, ctr, face, t
 
 (2) calculate timestep based on the CFL condition
 
