+++ docs/src/solver_update.md
@@ -1,0 +1,10 @@
+# Update
+
+```@docs
+KitBase.update!
+```
+
+The update solver calculate the variables at n+1 step based on numerical fluxes and in-cell collisions.
+- explicit
+- implicit
+- implicit-explicit (IMEX)
