+++ docs/src/solver_post.md
@@ -1,0 +1,8 @@
+# Post-process
+
+```@docs
+KitBase.plot_line
+KitBase.write_jld
+```
+
+The post-process solver handles the simulation data and visualization.
