+++ docs/src/solver_pre.md
@@ -1,0 +1,10 @@
+# Pre-process
+
+```@docs
+KitBase.initialize
+```
+
+The pre-process solver initializes the simulation that returns solver set, control volumes, interfaces, and current time.
+It could be a new simulation or restart of an interrupted one.
+- new run: .txt / .cfg / .toml / etc.
+- restart: .jld2
