+++ docs/src/solver_timestep.md
@@ -1,0 +1,7 @@
+# Timestep
+
+```@docs
+KitBase.timestep
+```
+
+The timestep solver returns the time interval used for the upcoming solution loop based on the current variables.
