+++ docs/src/solver_post.md
@@ -1,8 +1,8 @@
 # Post-process
 
 ```@docs
-KitBase.plot_line
-KitBase.write_jld
+plot_line
+write_jld
 ```
 
 The post-process solver handles the simulation data and visualization.
