+++ docs/src/solver_update.md
@@ -1,7 +1,7 @@
 # Update
 
 ```@docs
-KitBase.update!
+update!
 ```
 
 The update solver calculate the variables at n+1 step based on numerical fluxes and in-cell collisions.
