+++ docs/src/solver_reconstruction.md
@@ -1,23 +1,23 @@
 # Reconstruction
 
 ```@docs
-KitBase.reconstruct!
+reconstruct!
 ```
 
 The reconstruction solver interpolates piecewise solutions with the desirable order of accuracy.
 The reconstruction stencils can be based on 2 or 3 cells
 ```@docs
-KitBase.reconstruct2
-KitBase.reconstruct2!
-KitBase.reconstruct3
-KitBase.reconstruct3!
+reconstruct2
+reconstruct2!
+reconstruct3
+reconstruct3!
 ```
 
 The available schemes are
 ```@docs
-KitBase.vanleer
-KitBase.minmod
-KitBase.superbee
-KitBase.vanalbaba
-KitBase.weno5
+vanleer
+minmod
+superbee
+vanalbaba
+weno5
 ```
