+++ docs/src/solver_flux.md
@@ -1,7 +1,7 @@
 # Evolution
 
 ```@docs
-KitBase.evolve!
+evolve!
 ```
 
 The evolution solver calculate the interface numerical fluxes based on two neighbor cells.
@@ -10,17 +10,17 @@ The evolution solver calculate the interface numerical fluxes based on two neigh
 
 The available flux solvers are
 ```@docs
-KitBase.flux_lax!
-KitBase.flux_hll!
-KitBase.flux_roe!
-KitBase.flux_gks
-KitBase.flux_gks!
-KitBase.flux_kfvs!
-KitBase.flux_kcu!
-KitBase.flux_ugks!
-KitBase.flux_boundary_maxwell!
-KitBase.flux_boundary_specular!
-KitBase.flux_em!
-KitBase.flux_emx!
-KitBase.flux_emy!
+flux_lax!
+flux_hll!
+flux_roe!
+flux_gks
+flux_gks!
+flux_kfvs!
+flux_kcu!
+flux_ugks!
+flux_boundary_maxwell!
+flux_boundary_specular!
+flux_em!
+flux_emx!
+flux_emy!
 ```
