+++ docs/src/api_io.md
@@ -2,6 +2,4 @@
 
 ```@docs
 read_dict
-write_jld
-plot_line
 ```
