+++ docs/src/api_reconstruct.md
@@ -1,13 +1,0 @@
-# Reconstruction
-
-```@docs
-vanleer
-minmod
-superbee
-vanalbaba
-reconstruct2
-reconstruct2!
-reconstruct3
-reconstruct3!
-weno5
-```
