+++ docs/src/api_math.md
@@ -10,6 +10,6 @@ central_diff!
 upwind_diff
 upwind_diff!
 unstruct_diff
-lgwt
-extract_last
+KitBase.lgwt
+KitBase.extract_last
 ```
