+++ docs/src/api_flux.md
@@ -1,16 +1,0 @@
-# Flux
-
-```@docs
-flux_gks
-flux_gks!
-flux_kfvs!
-flux_kcu!
-flux_ugks!
-flux_boundary_maxwell!
-flux_lax!
-flux_hll!
-flux_roe!
-flux_em!
-flux_emx!
-flux_emy!
-```
