+++ docs/src/solver.md
@@ -1,6 +1,6 @@
 # General
 
-The finite volume method (FVM) is employed in Kinetic. 
+Kinetic employs the finite volume method (FVM) for modeling and simulation. 
 The general solution algorithm can be conclude as follows, where both explicit and implicit methods are implemented.
 
 ![](./assets/solver_process.png)
