### 2021-01-18 #2942 @Kevin-Mattheus-Moerman
@vavrines can you clarify what the entire codebase is that should be considered/reviewed here? The package Kinetic.jl appears too minor on its own. If the core functionality provided requires KitBase, KitML and KitFort (and perhaps kineticpy), all of which you wrote to accompany Kinetic (if I understand this correctly), then these should perhaps all be reviewed altogether and seen as a single submission, can you confirm?
