### 2021-01-18 #2942 @jedbrown
If the capability has rough parity with Clawpack, then there's no question about it being within scope. I can't tell if that's the case -- the examples are surely much less developed and documented. From briefly browsing the docs, I don't see where to look for boundary conditions, the extent of its multi-dimensional capability, time integrators, etc. What is supported? How fast is it compared to mature implementations (e.g., Clawpack or OpenFOAM)? The paper's statement of need should explain the capabilities and why you might opt for this package versus others.

Finally, I think drawing attention to the text input format is counter-productive. One of the advantages of Julia is programmatically creating model configurations and differentiating output with respect to arbitrary inputs. The text format seems to steer away from such things.
