+++ docs/src/sciml.md
@@ -1,0 +1,205 @@
+# Scientific Machine Learning
+
+Machine learning is building its momentum in scientific computing.
+Given the nonlinear structure of differential and integral equations, it is promising to incorporate the universal function approximator from machine learning models into the governing equations and achieve the balance between efficiency and accuracy.
+In the following, we present a universal differential equation strategy to construct the neural network enhanced Boltzmann equation.
+The complicated fivefold integral operator is replaced by a combination of relaxation and neural models.
+It promises a completely differential structure and thus the neural ODE type training and computing becomes possible.
+The approach reduces the computational cost up to three orders of magnitude and preserves the perfect accuracy.
+The detailed theory and implementation can be found in [Tianbai Xiao and Martin Frank, Using neural networks to accelerate the solution of the Boltzmann equation](https://arxiv.org/pdf/2010.13649.pdf).
+
+First we load all the packages needed and set up the configurations.
+```julia
+using OrdinaryDiffEq, Flux, DiffEqFlux, Plots
+using KitBase
+import KitML
+
+# config
+begin
+    case = "homogeneous"
+    maxTime = 3
+    tlen = 16
+    u0 = -5
+    u1 = 5
+    nu = 80
+    nug = 0
+    v0 = -5
+    v1 = 5
+    nv = 28
+    nvg = 0
+    w0 = -5
+    w1 = 5
+    nw = 28
+    nwg = 0
+    vMeshType = "rectangle"
+    nm = 5
+    knudsen = 1
+    inK = 0
+    alpha = 1.0
+    omega = 0.5
+    nh = 8
+end
+```
+
+The dataset is produced by the fast spectral method, which solves the nonlinear Boltzmann integral with fast Fourier transformation.
+```julia
+# dataset
+begin
+    tspan = (0.0, maxTime)
+    tran = linspace(tspan[1], tspan[2], tlen)
+    γ = heat_capacity_ratio(inK, 3)
+    vSpace = VSpace3D(u0, u1, nu, v0, v1, nv, w0, w1, nw, vMeshType)
+
+    f0 =
+        Float32.(
+            0.5 * (1 / π)^1.5 .*
+            (exp.(-(vSpace.u .- 0.99) .^ 2) .+ exp.(-(vSpace.u .+ 0.99) .^ 2)) .*
+            exp.(-vSpace.v .^ 2) .* exp.(-vSpace.w .^ 2),
+        ) |> Array
+    prim0 =
+        conserve_prim(moments_conserve(f0, vSpace.u, vSpace.v, vSpace.w, vSpace.weights), γ)
+    M0 = Float32.(maxwellian(vSpace.u, vSpace.v, vSpace.w, prim0)) |> Array
+
+    mu_ref = ref_vhs_vis(knudsen, alpha, omega)
+    kn_bzm = hs_boltz_kn(mu_ref, 1.0)
+    τ0 = mu_ref * 2.0 * prim0[end]^(0.5) / prim0[1]
+
+    phi, psi, phipsi = kernel_mode(
+        nm,
+        vSpace.u1,
+        vSpace.v1,
+        vSpace.w1,
+        vSpace.du[1, 1, 1],
+        vSpace.dv[1, 1, 1],
+        vSpace.dw[1, 1, 1],
+        vSpace.nu,
+        vSpace.nv,
+        vSpace.nw,
+        alpha,
+    )
+
+    # Boltzmann
[…]
