### 2021-01-18 #2942 @vavrines
I huess the editors may be discussing this article. To avoid possible misunderstandings, I'll write a brief reply here and sorry for any interruptions.

I agree Clawpack is a much more matured software. The main difference are listed here:

- Clawpack is coded in Fortran and Python, while Kinetic.jl is using Julia and avoids two-language problem
- It pays considerable attention to the Boltzmann equation, which is missing in Clawpack or OpenFOAM
- It is closely coupled with scientific machine learning

I believe the text input makes sense. During the runtime, the texts are going to generate variables directly, which relies on the meta-programming property of Julia. Different types of text input (txt, toml, cfg, etc.) are supported, and it provides a convenient way to modify settings without changing source codes. It is worth mentioning that there is an equivalent way to do configuration in Julia scripts without text input. I omitted it in the paper to avoid redundancy.

The documentation isn't perfect at present and is being organized continuously.
