+++ docs/src/tutorial.md
@@ -1,8 +1,13 @@
 # Tutorial 
 
-## Simulating gas dynamics in a few lines
+Thanks to the brilliant expressiveness and low-overhead abstraction, we provide different levels of solution algorithm for modeling and simulating advection-diffusion dynamics.
+The high-level solver is able to solve complex physics in a few lines, while the low-level APIs keep all the detailed implementations and are easy to be called from Python and C.
+In the following, we present two quick tutorials to illustrate the usage of Kinetic.
 
-In the following, we present an illustrative example of the shock tube problem in gas dynamics. 
+## Shock tube problem
+
+We first solve the shock tube problem in gas dynamics.
+It's a two dimensional problem, with one in physical domain ``x`` and another in particle velocity domain ``u``.
 First let us prepare the configuration file as
 ```
 # case
@@ -41,7 +46,6 @@ alphaRef = 1.0
 omegaRef = 0.5
 ```
 
-
 The configuration file can be understood as follows:
 - The simulation case is the standard Sod shock tube
 - A phase space in 1D physical and 1D velocity space is created with two particle distribution functions inside
@@ -64,8 +68,8 @@ The configuration file can be understood as follows:
 \mu_{ref}=\frac{5(\alpha+1)(\alpha+2) \sqrt{\pi}}{4 \alpha(5-2 \omega)(7-2 \omega)} Kn_{ref}
 ```
 
-
-The configuration can be stored in any text format. 
+The configuration file directly generate variables during runtime via meta-programming in Julia,
+and it can be stored in any text format (txt, toml, cfg, etc.). 
 For example, if `config.txt` is created, 
 we then execute the following codes to conduct a simulation
 ```julia
@@ -87,9 +91,269 @@ for iter = 1:nt
 end
 ```
 
-
 The result can be visualized with built-in function `plot_line`, which presents the profiles of gas density, velocity and temperature inside the tube.
 ```julia
 plot_line(set, ctr)
 ```
 ![](./assets/sod.png)
+
+
+## Lid-driven cavity
+
+We then show the lid-driven cavity.
+It's a four dimensional problem, with two in physical domain ``(x,y)`` and another in particle velocity domain ``(u,v)``.
+Similarly, we prepare the configuration file as
+```
+# setup
+case = cavity
+space = 2d2f2v
+flux = kfvs
+collision = bgk
+nSpecies = 1
+interpOrder = 2
+limiter = vanleer
+boundary = maxwell
+cfl = 0.8
+maxTime = 2.0
+
+# phase space
+x0 = 0.0
+x1 = 1.0
+nx = 45
+y0 = 0.0
+y1 = 1.0
+ny = 45
+pMeshType = uniform
+nxg = 0
+nyg = 0
+
+# velocity space
+umin = -5.0
+umax = 5.0
+nu = 28
+vmin = -5.0
+vmax = 5.0
+nv = 28
+vMeshType = rectangle
+nug = 0
+nvg = 0
+
+# gas
+knudsen = 0.075
+mach = 2.0
+prandtl = 1.0
+inK = 1.0
+omega = 0.72
+alphaRef = 1.0
+omegaRef = 0.5
+
+# boundary
+uLid = 0.15
+vLid = 0.0
+tLid = 1.0
+```
+
+We then execute the following codes to conduct a simulation
+```julia
+using Kinetic
+set, ctr, a1face, a2face, t = initialize("config.txt")
+t = solve!(set, ctr, a1face, a2face, t)
+```
+
+The high-level solver `solve!` is equivalent as the following low-level procedures
+```julia
[…]
