### 2021-01-19 #2942 @kthyng
@vavrines Can I verify that this submission covers Kinetic.jl and KitBase, KitML and KitFort? Also, have these other packages been published on or are in the process of being published on?

We seem to be narrowing in on what exactly is in this submission and that the packages together are in scope. However, there is still concern about the maturity of the project at this point. The paper is incomplete with no statement of need and a reference to a figure that is not present (?) along with other errors. Also, given that as part of the JOSS review, two reviewers will go through the package(s) in detail as well as all of the documentation, these items should be as perfect as possible now as opposed to some point in the future. Tests and other [requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) should also be fully functional before the review starts.

Would you like to pause this submission to improve the documentation and paper before proceeding? This would mean you think you could improve it enough in, say, a few weeks, to be able to proceed with the review. Alternatively, you could withdraw and resubmit at a later time when you have had a chance to make these improvements.
