### 2021-01-19 #2942 @vavrines
> @vavrines Can I verify that this submission covers Kinetic.jl and KitBase, KitML and KitFort? Also, have these other packages been published on or are in the process of being published on?
>
> We seem to be narrowing in on what exactly is in this submission and that the packages together are in scope. However, there is still concern about the maturity of the project at this point. The paper is incomplete with no statement of need and a reference to a figure that is not present (?) along with other errors. Also, given that as part of the JOSS review, two reviewers will go through the package(s) in detail as well as all of the documentation, these items should be as perfect as possible now as opposed to some point in the future. Tests and other [requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) should also be fully functional before the review starts.
>
> Would you like to pause this submission to improve the documentation and paper before proceeding? This would mean you think you could improve it enough in, say, a few weeks, to be able to proceed with the review. Alternatively, you could withdraw and resubmit at a later time when you have had a chance to make these improvements.

Dear @kthyng
Thank you very much for the reply. I can confirm this submission covers Kinetic.jl, KitBase.jl, KitML.jl and KitFort.jl. All these packages have neber been published on or are in process of being published on any other journal.

And thanks for the comments on the paper. I don't agree with part of the arguments. The statement of need is obviously provided in the summary of paper. The figure is tightly connected to the text above, and I can add a caption of it if necessary (not sure about this point in JOSS). The tests and other requiements work well.

I think the current maturity of the documentation and of the whole package meets the readership of JOSS. However, if it comes to an official decision from editors that the paper submission has to be paused, I would respect it and apply myself to the improvement.

Best,
Tianbai
