### 2021-01-19 #2942 @kthyng
> The statement of need is obviously provided in the summary of paper.

We require a section entitled "Statement of Need." @jedbrown gave suggestions for additions above.

> The figure is tightly connected to the text above, and I can add a caption of it if necessary (not sure about this point in JOSS).

The paper seems to imply the presence of a figure but I do not see one. Am I missing something? I see it now in the github representation of the paper but not in the whedon compiled version. Yes you can and should include a caption. The information you need to implement these items should be [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

> The tests and other requiements work well.

At a minimum, you can see @jedbrown's comments about the state of the documentation, but this came up in the editorial discussion outside of this issue as well.

The low coverage of tests were brought up as well. Can you elaborate on what is tested for at present?

The items that I am bringing up now are from the editorial board discussion and are also the items that reviewers would need to bring up immediately upon starting their reviews so we try to catch things like this before the review process starts when possible to reduce the time reviewers need to spend. I will pause this paper for now as you work through the items.
