### 2021-01-19 #2942 @vavrines
@kthyng
Thanks again for the kind explanation.

I figured out that JOSS seems not accepting html syntax so that all the figures in (https://github.com/vavrines/Kinetic.jl/blob/master/paper/paper.md) are missing. Is there a recommended way to adjust the attributes of a picture in JOSS?

I'll work on improving the paper. When it's ready, how should I contact the editor and restart the review process?

Best, Tianbai
