### 2021-01-19 #2942 @kthyng
@vavrines can you use the example given in the docs that is like this: `![Caption for example figure.](figure.png){ width=20% }`

In addition to the paper, please work on the docs and tests.

When you are ready, you can ping @openjournals/joss-eics  to start your submission back up.
