+++ Project.toml
@@ -4,22 +4,21 @@ authors = ["Tianbai Xiao <tianbaixiao@gmail.com>"]
 version = "0.7.0"
 
 [deps]
-Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
 KitBase = "86eed249-3a28-466f-8d3a-596821e1af9a"
 KitML = "b2531deb-9288-45c6-8f4a-0220551da45d"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
-Conda = "^1"
 KitBase = "0.3"
 KitML = "0.1"
-PyCall = "^1"
 Reexport = "0.2, 1.0"
 julia = "^1"
 
 [extras]
+Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
+PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
+build = ["PyCall", "Conda"]
 test = ["Test"]
