+++ docs/make.jl
@@ -1,9 +1,23 @@
-#push!(LOAD_PATH,"../src/")
 import Pkg
 Pkg.add("Documenter")
 
 using Documenter, Kinetic
 using Kinetic: KitBase, KitML
+
+tutorial_page = [
+    "Shock tube problem" => "eg_shock.md",
+    "Lid-driven cavity" => "eg_cavity.md",
+]
+
+solver_page = [
+    "General" => "solver.md",
+    "Preprocess" => "solver_pre.md",
+    "Timestep" => "solver_timestep.md",
+    "Reconstruction" => "solver_reconstruction.md",
+    "Flux" => "solver_flux.md",
+    "Update" => "solver_update.md",
+    "Postprocess" => "solver_post.md",
+]
 
 makedocs(
     sitename= "Kinetic.jl",
@@ -11,20 +25,11 @@ makedocs(
     pages = Any[
         "Home" => "index.md",
         "Installation" => "install.md",
-        "Tutorial" => "tutorial.md",
+        "Tutorial" => tutorial_page,
         "Physics" => "physics.md",
         "Type" => "type.md",
-        "Solver" => [
-            "solver.md",
-            "solver_pre.md",
-            "solver_timestep.md",
-            "solver_reconstruction.md",
-            "solver_flux.md",
-            "solver_update.md",
-            "solver_post.md",
-            ],
-        "SciML" => "sciml.md",
-        "Auxiliary" => [
+        "Solver" => solver_page,
+        "Utility" => [
             "api_io.md",
             "api_math.md",
             "api_theory.md",
@@ -33,6 +38,9 @@ makedocs(
             "api_config.md",
             "api_step.md",
             ],
+        "KitML" => "kitml.md",
+        "KitFort" => "kitfort.md"
+        "Index" => "function_index.md"
     ]
 )
 
