+++ docs/src/eg_cavity.md
@@ -1,0 +1,259 @@
+# Lid-Driven Cavity
+
+We then show the lid-driven cavity.
+It's a four dimensional problem, with two in physical domain ``(x,y)`` and another in particle velocity domain ``(u,v)``.
+Similarly, we prepare the configuration file as
+```
+# setup
+case = cavity
+space = 2d2f2v
+flux = kfvs
+collision = bgk
+nSpecies = 1
+interpOrder = 2
+limiter = vanleer
+boundary = maxwell
+cfl = 0.8
+maxTime = 2.0
+
+# phase space
+x0 = 0.0
+x1 = 1.0
+nx = 45
+y0 = 0.0
+y1 = 1.0
+ny = 45
+pMeshType = uniform
+nxg = 0
+nyg = 0
+
+# velocity space
+umin = -5.0
+umax = 5.0
+nu = 28
+vmin = -5.0
+vmax = 5.0
+nv = 28
+vMeshType = rectangle
+nug = 0
+nvg = 0
+
+# gas
+knudsen = 0.075
+mach = 2.0
+prandtl = 1.0
+inK = 1.0
+omega = 0.72
+alphaRef = 1.0
+omegaRef = 0.5
+
+# boundary
+uLid = 0.15
+vLid = 0.0
+tLid = 1.0
+```
+
+We then execute the following codes to conduct a simulation
+```julia
+using Kinetic
+set, ctr, a1face, a2face, t = initialize("config.txt")
+t = solve!(set, ctr, a1face, a2face, t)
+```
+
+The high-level solver `solve!` is equivalent as the following low-level procedures
+```julia
+using ProgressMeter
+res = zeros(4)
+dt = timestep(ks, ctr, simTime)
+nt = floor(ks.set.maxTime / dt) |> Int
+@showprogress for iter = 1:nt
+    reconstruct!(ks, ctr)
+    evolve!(ks, ctr, a1face, a2face, dt; mode = Symbol(ks.set.flux), bc = Symbol(ks.set.boundary))
+    update!(ks, ctr, a1face, a2face, dt, res; coll = Symbol(ks.set.collision), bc = Symbol(ks.set.boundary))
+end
+```
+
+It can be further expanded into the lower-level backend.
+```julia
+# lower-level backend 
+@showprogress for iter = 1:nt
+    # horizontal flux
+    @inbounds Threads.@threads for j = 1:ks.pSpace.ny
+        for i = 2:ks.pSpace.nx
+            KitBase.flux_kfvs!(
+                a1face[i, j].fw,
+                a1face[i, j].fh,
+                a1face[i, j].fb,
+                ctr[i-1, j].h,
+                ctr[i-1, j].b,
+                ctr[i, j].h,
+                ctr[i, j].b,
+                ks.vSpace.u,
+                ks.vSpace.v,
+                ks.vSpace.weights,
+                dt,
+                a1face[i, j].len,
+            )
+        end
+    end
+    
+    # vertical flux
+    vn = ks.vSpace.v
+    vt = -ks.vSpace.u
+    @inbounds Threads.@threads for j = 2:ks.pSpace.ny
+        for i = 1:ks.pSpace.nx
+            KitBase.flux_kfvs!(
+                a2face[i, j].fw,
+                a2face[i, j].fh,
+                a2face[i, j].fb,
+                ctr[i, j-1].h,
+                ctr[i, j-1].b,
+                ctr[i, j].h,
+                ctr[i, j].b,
+                vn,
+                vt,
+                ks.vSpace.weights,
+                dt,
+                a2face[i, j].len,
+            )
+            a2face[i, j].fw .= KitBase.global_frame(a2face[i, j].fw, 0., 1.)
+        end
+    end
+    
+    # boundary flux
[…]
