+++ README.md
@@ -33,6 +33,21 @@ and continuum mechanics, e.g. the Euler and Navier-Stokes equations
 
 <a href="https://www.codecogs.com/eqnedit.php?latex=\frac{\partial&space;\mathbf&space;W}{\partial&space;t}&space;&plus;&space;\nabla_\mathbf&space;x&space;\cdot&space;\mathbf&space;F&space;=&space;\mathbf&space;S" target="_blank"><img src="https://latex.codecogs.com/svg.latex?\frac{\partial&space;\mathbf&space;W}{\partial&space;t}&space;&plus;&space;\nabla_\mathbf&space;x&space;\cdot&space;\mathbf&space;F&space;=&space;\mathbf&space;S" title="\frac{\partial \mathbf W}{\partial t} + \nabla_\mathbf x \cdot \mathbf F = \mathbf S" /></a>
 
+A list of current supported equations include:
+- linear Boltzmann equation
+- nonlinear Boltzmann equation
+- nonlinear kinetic model equation
+- multi-component Boltzmann equations
+- advection-diffusion equation
+- Burgers equation
+- Euler equations
+- Navier-Stokes equations
+- Extended hydrodynamical equations from gas kinetic expansion
+- Magnetohydrodynamical equations
+- Poisson equation
+- Fokker-Planck-Landau equation
+- Maxwell's equations
+
 ## Documentation
 
 For the detailed information on the implementation and usage of the package, please
