+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"solver_update/#Update","page":"Update","title":"Update","text":"","category":"section"},{"location":"solver_update/","page":"Update","title":"Update","text":"update!","category":"page"},{"location":"solver_update/#KitBase.update!","page":"Update","title":"KitBase.update!","text":"update!(\n    KS::X,\n    ctr::Y,\n    face::Z,\n    dt,\n    residual; # 1D / 2D\n    coll = :bgk::Symbol,\n    bc = :fix::Symbol,\n) where {\n    X<:AbstractSolverSet,\n    Y<:AbstractArray{ControlVolume1D1F,1},\n    Z<:AbstractArray{Interface1D1F,1},\n}\n\nupdate!(\n    KS::X,\n    ctr::Y,\n    face::Z,\n    dt,\n    residual; # 1D / 2D\n    coll = :bgk::Symbol,\n    bc = :extra::Symbol,\n) where {\n    X<:AbstractSolverSet,\n    Y<:AbstractArray{ControlVolume1D2F,1},\n    Z<:AbstractArray{Interface1D2F,1},\n}\n\nupdate!(\n    KS::X,\n    ctr::Y,\n    face::Z,\n    dt,\n    residual; # 1D / 2D\n    coll = :bgk::Symbol,\n    bc = :extra::Symbol,\n    isMHD = true::Bool,\n) where {\n    X<:AbstractSolverSet,\n    Y<:AbstractArray{ControlVolume1D3F,1},\n    Z<:AbstractArray{Interface1D3F,1},\n}\n\nupdate!(\n    KS::X,\n    ctr::Y,\n    face::Z,\n    dt,\n    residual; # 1D / 2D\n    coll = :bgk::Symbol,\n    bc = :extra::Symbol,\n    isMHD = true::Bool,\n) where {\n    X<:AbstractSolverSet,\n    Y<:AbstractArray{ControlVolume1D4F,1},\n    Z<:AbstractArray{Interface1D4F,1},\n}\n\nupdate!(\n    KS::X,\n    ctr::Y,\n    a1face::Z,\n    a2face::Z,\n    dt,\n    residual; # 1D / 2D\n    coll = :bgk::Symbol,\n    bc = :fix::Symbol,\n) where {\n    X<:AbstractSolverSet,\n    Y<:AbstractArray{ControlVolume2D2F,2},\n    Z<:AbstractArray{Interface2D2F,2},\n}\n\nUpdate flow variables over control volumes\n\n\n\n\n\n","category":"function"},{"location":"solver_update/","page":"Update","title":"Update","text":"The update solver calculate the variables at n+1 step based on numerical fluxes and in-cell collisions.","category":"page"},{"location":"solver_update/","page":"Update","title":"Update","text":"explicit\nimplicit\nimplicit-explicit (IMEX)","category":"page"},{"location":"api_theory/#Theory","page":"Theory","title":"Theory","text":"","category":"section"},{"location":"api_theory/","page":"Theory","title":"Theory","text":"prim_conserve\nconserve_prim\nmixture_prim_conserve\nmixture_conserve_prim\nem_coefficients\nadvection_flux\nburgers_flux\neuler_flux\neuler_jacobi\ngauss_moments\nmixture_gauss_moments\nmoments_conserve\nmixture_moments_conserve\npdf_slope\nmixture_pdf_slope\nmoments_conserve_slope\nmixture_moments_conserve_slope\ndiscrete_moments\nstress\nheat_flux\nmaxwellian\nmixture_maxwellian\nshakhov\nreduce_distribution\nfull_distribution\n
[…]
