+++ docs/src/api_step.md
@@ -3,16 +3,3 @@
 ```@docs
 KitBase.step!
 ```
-
-
-Kinetic.initialize
-Kinetic.init_fvm
-SolverSet
-set_ib
-solve!
-Kinetic.timestep
-reconstruct!
-evolve!
-update!
-update_boundary!
-step!
