+++ .github/workflows/ci.yml
@@ -15,7 +15,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.5' # Replace this with the minimum Julia version that your package supports. E.g. if your package requires Julia 1.5 or higher, change this to '1.5'.
+          - '1.5' # Replace this with the minimum Julia version that the package supports.
+          - '1.6'
           - '1' # Leave this line unchanged. '1' will automatically expand to the latest stable 1.x release of Julia.
 #          - 'nightly'
         os:
