+++ .github/workflows/ci.yml
@@ -16,7 +16,6 @@ jobs:
       matrix:
         version:
           - '1.5' # Replace this with the minimum Julia version that the package supports.
-          - '1.6'
           - '1' # Leave this line unchanged. '1' will automatically expand to the latest stable 1.x release of Julia.
 #          - 'nightly'
         os:
