+++ docs/Project.toml
@@ -2,6 +2,8 @@
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
+KitFort = "c7785b10-740d-4428-9d22-c85dcf833c60"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
 
 [compat]
