+++ docs/make.jl
@@ -4,7 +4,6 @@ Pkg.add("KitFort")
 
 using Documenter, Kinetic, KitFort
 using Kinetic: KitBase, KitML
-using Reexport
 
 tutorial_page = [
     "Shock tube problem" => "eg_shock.md",
