+++ docs/src/kitml.md
@@ -1,4 +1,4 @@
-# KitML and Scientific Machine Learning
+# Scientific Machine Learning and KitML
 
 Machine learning is building its momentum in scientific computing.
 Given the nonlinear structure of differential and integral equations, it is promising to incorporate the universal function approximator from machine learning models into the governing equations and achieve the balance between efficiency and accuracy.
@@ -11,8 +11,7 @@ The detailed theory and implementation can be found in [Tianbai Xiao and Martin 
 First we load all the packages needed and set up the configurations.
 ```julia
 using OrdinaryDiffEq, Flux, DiffEqFlux, Plots
-using KitBase
-import KitML
+using KitBase, KitML
 
 # config
 begin
