+++ docs/src/function_index.md
@@ -1,4 +1,4 @@
-# Function Index
+# Index of Types and Methods
 
 ```@index
 ```
