### 2021-01-22 05c3811
Update docs of type
