+++ docs/src/index.md
@@ -1,27 +1,50 @@
 # Kinetic.jl
 
-Kinetic is a portable Julia toolbox for the study of computational fluid dynamics and scientific machine learning.
-The main module consists of [KitBase.jl](https://github.com/vavrines/KitBase.jl) with basic physics and [KitML.jl](https://github.com/vavrines/KitML.jl) with neural dynamics. 
-The high-performance Fortran library [KitFort.jl](https://github.com/vavrines/KitFort.jl) is optional and can be manually imported when the executing efficiency becomes the first priority. 
+Kinetic is a portable toolbox for the study of computational fluid dynamics and scientific machine learning.
+The default module consists of [KitBase.jl](https://github.com/vavrines/KitBase.jl) with basic physics and [KitML.jl](https://github.com/vavrines/KitML.jl) with neural dynamics. 
+The high-performance Fortran library [KitFort.jl](https://github.com/vavrines/KitFort.jl) can be manually imported when the executing efficiency becomes the first priority. 
 A wrapper [kineticpy](https://github.com/vavrines/kineticpy) has been built to locate the data hierarchies and methods in Python.
 
-## Topics
+## Scope of application
 
-Kinetic is interested in theoretical modeling and numerical simulation of many-particle systems, e.g. gases, photons, plasmas, neutrons, electrons, etc.
-It employs the finite volume method (FVM) to conduct 1-3 dimensional numerical simulations on CPUs and GPUs.
+Kinetic is interested in the evolution of many-particle systems, e.g. gases, photons, plasmas, neutrons, electrons, etc.
+Based on the finite volume method (FVM), it provides a direct modeling tool in the discrete space,
+where 1-3 dimensional theoretical modeling and numerical simulation can be conducted.
 Any advection-diffusion type equation can be hooked and solved within the framework.
-The user interface is designed as intuitive and neat as possible.
-The combined development of high-level and low-level APIs allows users to focus on physics and to cooperate with existing packages in the Julia ecosystem.
+Special attention has been paid to the [kinetic theory](https://en.wikipedia.org/wiki/Kinetic_theory_of_gases) and the Boltzmann-type equations,
+which depicts the time-space evolution of particles via ensemble averaging at the mesoscopic level.
+A partial list of current supported models and equations include:
+- linear Boltzmann equation
+- nonlinear Boltzmann equation
+- multi-component Boltzmann equations
+- Fokker-Planck-Landau equation
+- direct simulation Monte Carlo
+- advection-diffusion equation
+- Burgers equation
+- Euler equations
+- Navier-Stokes equations
+- Extended hydrodynamical equations from gas kinetic expansion
+- Magnetohydrodynamical equations
+- Maxwell's equations
 
-## Novelty
+## Design philosophy
+
+The code hierarchy is designed as intuitive and neat as possible.
+It's dedicated to providing a friendly interface for educational usage in kinetic theory and rich functionality for scientific research.
+Benefiting from the brilliant expressiveness and low-overhead abstraction provided by the [Julia programming language](https://julialang.org/), 
+we provide a combination of high-level and low-level APIs to allow the users to focus on physics and to cooperate with the existing packages in the Julia ecosystem.
+
+## What is new?
 
 Finite volume method is a proven approach for simulating conservation laws.
-Compared with the existing open-source softwares, e.g. OpenFOAM, SU2 and Clawpack, Kinetic holds the following innovative points
-- 100% Julia stack with no two-language problem
-- Well-established support for the Boltzmann and other integro-differential equations in phase space
-- Closely coupled with scientific machine learning
+Compared with the existing open-source softwares, e.g. [OpenFOAM](https://openfoam.org/), [SU2](https://su2code.github.io/) and [Clawpack](https://www.clawpack.org/), 
[…]
