+++ docs/src/tutorial.md
@@ -1,0 +1,7 @@
+# Illustrative examples
+
+Thanks to the brilliant expressiveness and low-overhead abstraction in Julia, we provide different levels of solution algorithm for modeling and simulating advection-diffusion dynamics.
+The high-level solver is able to solve complex physics in a few lines, while the low-level APIs keep all the detailed implementations and benefit the secondary development.
+The low-level methods are easy to be called from Python and C.
+In the following, we present some quick tutorials to illustrate the usage of Kinetic.
+For more examples, please refer the example directories in [Kinetic.jl](https://github.com/vavrines/Kinetic.jl/tree/master/example), [KitBase.jl](https://github.com/vavrines/KitBase.jl/tree/main/example) and [KitML.jl](https://github.com/vavrines/KitML.jl/tree/main/example).
