+++ benchmark/bench_ds.txt
@@ -10,6 +10,7 @@ flux = kfvs
 collision = bgk
 interpOrder = 2
 limiter = vanleer
+boundary = fix
 cfl = 0.5
 maxTime = 0.2
 
