+++ docs/src/type_2d.md
@@ -1,6 +1,6 @@
-# Two-dimensional data
+# Two-dimensional FVM
 
-The 2D control volume structs are
+The 2D control volume structs are implemented.
 ```@docs
 ControlVolume2D
 ControlVolume2D1F
@@ -8,19 +8,9 @@ ControlVolume2D2F
 ControlVolume2D3F
 ```
 
-The numerical fluxes are evaluated through `AbstractInterface` structs
+The numerical fluxes are evaluated through `AbstractInterface` structs.
 ```@docs
 Interface2D
 Interface2D1F
 Interface2D2F
-```
-
-The rest structs for saving general computational setups are
-```@docs
-SolverSet
-```
-
-```@autodocs
-Modules = [KitBase]
-Order = [:type]
 ```
