+++ docs/src/api_geo.md
@@ -3,12 +3,8 @@
 ```@docs
 global_frame
 local_frame
-PSpace1D
-PSpace2D
 uniform_mesh
 meshgrid
-UnstructMesh 
-read_mesh
 mesh_connectivity_2D
 mesh_center_2D
 mesh_area_2D
