+++ docs/src/api_phase.md
@@ -1,12 +1,6 @@
 # Phase Space
 
 ```@docs
-VSpace1D
-VSpace2D
-VSpace3D
-MVSpace1D
-MVSpace2D
-MVSpace3D
 newton_cotes
 legendre_quadrature
 octa_quadrature
