+++ docs/src/type_2d.md
@@ -1,16 +1,0 @@
-# Two-dimensional FVM
-
-The 2D control volume structs are implemented.
-```@docs
-ControlVolume2D
-ControlVolume2D1F
-ControlVolume2D2F
-ControlVolume2D3F
-```
-
-The numerical fluxes are evaluated through `AbstractInterface` structs.
-```@docs
-Interface2D
-Interface2D1F
-Interface2D2F
-```
