+++ docs/src/solver_post.md
@@ -1,8 +1,9 @@
-# Post-process
+# Postprocess
+
+The post-process solver handles the simulation data and visualization.
 
 ```@docs
 plot_line
+plot_contour
 write_jld
 ```
-
-The post-process solver handles the simulation data and visualization.
