+++ docs/src/para_cuda.md
@@ -1,0 +1,7 @@
+# GPU computing
+
+```julia
+CUDA
+```
+
+architecture
