+++ docs/Project.toml
@@ -1,3 +1,10 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 Kinetic = "82403725-3cee-4f7c-b214-1ce71af4a797"
+StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
+
+[compat]
+Documenter = "0.25"
+DocumenterTools = "0.1"
