+++ paper/paper.bib
@@ -1,3 +1,15 @@
+@book{chapman1990,
+  title={The mathematical theory of non-uniform gases: an account of the kinetic theory of viscosity, thermal conduction and diffusion in gases},
+  author={Chapman, Sydney and Cowling, Thomas George and Burnett, David},
+  year={1990},
+  publisher={Cambridge university press}
+}
+@book{batchelor2000,
+  title={An introduction to fluid dynamics},
+  author={Batchelor, Cx K and Batchelor, GK},
+  year={2000},
+  publisher={Cambridge university press}
+}
 @article{bezanson2017,
   title={Julia: A fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
@@ -18,12 +30,12 @@
   organization={IUC Dubrovnik Croatia}
 }
 @misc{clawpack2020,
-    title={Clawpack software},
-    author={{Clawpack Development Team}},
-    url={http://www.clawpack.org},
-    note={Version 5.7.1},
-    doi={https://doi.org/10.5281/zenodo.4025432},
-    year={2020}}
+  title={Clawpack software},
+  author={{Clawpack Development Team}},
+  url={http://www.clawpack.org},
+  note={Version 5.7.1},
+  doi={https://doi.org/10.5281/zenodo.4025432},
+  year={2020}}
 @article{Flux2018,
   author    = {Michael Innes and
                Elliot Saba and
@@ -46,10 +58,19 @@
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 @misc{xiao2020,
-      title={Using neural networks to accelerate the solution of the Boltzmann equation}, 
-      author={Tianbai Xiao and Martin Frank},
-      year={2020},
-      eprint={2010.13649},
-      archivePrefix={arXiv},
-      primaryClass={physics.comp-ph}
+  title={Using neural networks to accelerate the solution of the Boltzmann equation}, 
+  author={Tianbai Xiao and Martin Frank},
+  year={2020},
+  eprint={2010.13649},
+  archivePrefix={arXiv},
+  primaryClass={physics.comp-ph}
+}
+@article{xiao2017,
+  title={A well-balanced unified gas-kinetic scheme for multiscale flow transport under gravitational field},
+  author={Xiao, Tianbai and Cai, Qingdong and Xu, Kun},
+  journal={Journal of Computational Physics},
+  volume={332},
+  pages={475--491},
+  year={2017},
+  publisher={Elsevier}
 }
