+++ paper/paper.md
@@ -1,8 +1,8 @@
 ---
-title: 'Kinetic.jl: A lightweight finite volume toolbox in Julia'
+title: 'Kinetic.jl: A portable finite volume toolbox in Julia'
 tags:
-  - computational fluid dynamics
   - kinetic theory
+  - computational fluid dynamics
   - scientific machine learning
   - julia
 authors:
@@ -18,91 +18,130 @@ bibliography: paper.bib
 
 # Summary
 
-``Kinetic.jl`` is a lightweight toolbox for computational fluid dynamics and scientific machine learning. It focus on the theoretical and numerical studies of many-particle systems of gases, photons, plasmas, neutrons, etc. Under the hood, the finite volume method (FVM) is implemented with various flux functions and update algorithms. Therefore, 1-3 dimensional numerical simulations can be conducted for any advection-diffusion type equations. Special attentions have been paid on Hilbert's sixth problem, i.e. to build the numerical passage between kinetic theory of gases, e.g. the Boltzmann equation
+Kinetic.jl is a lightweight finite volume toolbox written in the Julia programming language for the study of computational mechanics and scientific machine learning.
+It is an open-source project hosted on GitHub and distributed under MIT license.
+The main module consists of KitBase.jl with basic physics and KitML.jl with neural dynamics.
+The library provides a rich set of numerical flux functions.
+Any advection-diffusion type mechanical or neural equation can be hooked and solved in the framework.
+It is designed to balance the programming flexibility for scientific research, the algorithmic efficiency for application, and the simplicity for educational usage.
 
-<img src="bz.png" width = "600"  align=center />
+# Statement of need
 
-and the Navier-Stokes equations.
+A physical system can perform a wonderfully diverse set of acts on different characteristic scales.
+For example, particle transports can be described by kinetic theory of gases at particle mean free path scale [@chapman1990] and fluid mechanics at macroscopic level [@batchelor2000].
+With the rapidly advancing computing power, the finite volume method (FVM) is a prevalent method for quantitatively describing physical evolutions.
 
-<img src="ns.png" width = "150"  align=center />
+Most existing FVM libraries, e.g. the OpenFOAM [@jasak2007], are written in compiled languages (C/C++ and Fortran), which enjoy the perfect execution efficiency but sacrifice the flexibility for development.
+For example, it is cumbersome to implement the phase-field evolution from the Boltzmann equation [@xiao2017] in OpenFOAM or integrate it with scientific machine learning (SciML) packages.
+One compromise can be made with a combination of static and dynamic languages [@clawpack2020], where the high-level front-ends and the low-level computational back-ends are splitted.
+Basically it benefits general users, while researchers still need to work on the back-end if a new feature is required. 
+Besides, the two-language problem introduces additional trade-off in both development and execution.
+Different from these packages, Kinetic.jl is built upon the Julia programming language [@bezanson2017], which is dynamically typed and designed for high performance computing for broad devices. 
+Based on type inference and multiple dispatch, it is a promising choice to solve the two-language problem.
 
[…]
