+++ docs/src/fortran1.md
@@ -1,4 +1,4 @@
-# KitFort.jl
+# KitFort and high performance computing
 
 Numerical simulations of nonlinear models and differential equations are essentially connected with supercomputers and high-performance computing (HPC).
 The performance of a supercomputer or a software program is commonly measured in floating-point operations per second (FLOPS).
