### 2021-02-04 #2942 @vavrines
With the revise of the paper, docs and tests, I would like to reboot the review process.

The major changes include:
- The paper is rewritten. A statement of need section is added. The figure format is adjusted. The proof reading is made.
- The tests are expanded and the coverage increases.
- The documentation is reorganized.

@openjournals/joss-eics
@arfon
@kthyng
@Kevin-Mattheus-Moerman
Could you please help continue the review?
