+++ v0.7.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"para_dist/#Distributed-computing","page":"Distributed","title":"Distributed computing","text":"","category":"section"},{"location":"para_dist/","page":"Distributed","title":"Distributed","text":"The distributed computation is built upon Julia's @distributed macro in the Distributed module.","category":"page"},{"location":"para_dist/","page":"Distributed","title":"Distributed","text":"using Distributed\n@distributed [reducer] for var = range\n    body\nend","category":"page"},{"location":"para_dist/","page":"Distributed","title":"Distributed","text":"It provides a MPI-type parallelization with a leaner code size. The specified range is partitioned and locally executed across all workers.  In case an optional reducer function is specified, @distributed performs local reductions on each worker with a final reduction on the calling process. Without a reducer function, @distributed will execute asynchronously, i.e. it spawns independent tasks on all available workers and returns immediately without waiting for completion.  To make it wait for completion, prefix the call with @sync like :","category":"page"},{"location":"para_dist/","page":"Distributed","title":"Distributed","text":"@sync @distributed for var = range\n    body\nend","category":"page"},{"location":"para_dist/","page":"Distributed","title":"Distributed","text":"In the following, we present an example to conduct distributed computing with the help Julia's SharedArrays module, which creates arrays shared by all the processors. More massive computing can be made with DistributedArrays. First, we consider a distributed computing.","category":"page"},{"location":"para_dist/","page":"Distributed","title":"Distributed","text":"using Distributed, SharedArrays\naddprocs(3)\n@everywhere using KitBase\n\nbegin\n    vars = Dict{Symbol,Any}()\n    vars[:matter] = \"gas\"\n    vars[:case] = \"sod\"\n    vars[:space] = \"1d0f0v\"\n    vars[:flux] = \"kfvs\"\n    vars[:collision] = \"bgk\"\n    vars[:nSpecies] = 1\n    vars[:interpOrder] = 1\n    vars[:limiter] = \"vanleer\"\n    vars[:boundary] = \"fix\"\n    vars[:cfl] = 0.5\n    vars[:maxTime] = 0.2\n    vars[:x0] = 0.0\n    vars[:x1] = 1.0\n    vars[:nx] = 2000\n    vars[:pMeshType] = \"uniform\"\n    vars[:nxg] = 0\n    vars[:knudsen] = 0.001\n    vars[:mach] = 0.0\n    vars[:prandtl] = 1.0\n    vars[:inK] = 0.0\n    vars[:omega] = 0.81\n    vars[:alphaRef] = 1.0\n    vars[:omegaRef] = 0.5\nend\n\nset = KitBase.set_setup(vars)\npSpace = KitBase.set_geometry(vars)\nvSpace = KitBase.set_velocity(vars)\ngas = KitBase.set_property(vars)\nib = KitBase.set_ib(vars, set, vSpace, gas)\nfolder = @__DIR__\nks = KitBase.SolverSet(set, pSpace, vSpace, gas, ib, folder)\n\ndt = ks.pSpace.dx[1] / (5.0 + KitBase.sound_speed(ks.ib.primL, ks.gas.γ))\nnt = floor(ks.set.maxTim
[…]
