### 2021-02-11 #2942 @diehlpk
Hey @ChrisRackauckas, @rdeits, @dpsanders, and @KristofferC would you be interested to review this paper?
