### 2021-02-15 #2942 @rdeits
Sure, I can help out. I've done lots of scientific paper reviews in grad school, but never anything as software-focused as this, so I'm not totally sure what's expected. Do you have any info about what you're expecting from a JOSS reviewer?
