### 2021-02-16 #2942 @diehlpk
@rdeits Please have a look at our guidelines

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Please let me know if you like to have more details?
