### 2021-02-23 #3060 @whedon
[REVIEW]: Kinetic.jl: A lightweight finite volume toolbox in Julia
**Submitting author:** @vavrines (<a href="http://orcid.org/0000-0001-9127-9497">Tianbai Xiao</a>)
**Repository:** <a href="https://github.com/vavrines/Kinetic.jl" target ="_blank">https://github.com/vavrines/Kinetic.jl</a>
**Version:** v0.7.5
