+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Kinetic"
 uuid = "82403725-3cee-4f7c-b214-1ce71af4a797"
 authors = ["Tianbai Xiao <tianbaixiao@gmail.com>"]
-version = "0.7.1"
+version = "0.7.2"
 
 [deps]
 KitBase = "86eed249-3a28-466f-8d3a-596821e1af9a"
@@ -9,8 +9,8 @@ KitML = "b2531deb-9288-45c6-8f4a-0220551da45d"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
-KitBase = "0.3"
-KitML = "0.2"
+KitBase = "0.4"
+KitML = "0.3"
 Reexport = "0.2, 1.0"
 julia = "^1"
 
