### 2021-03-10 140b737
Add burgers example
