### 2021-03-24 ed334f5
Update type_domain.md
