### 2021-04-08 vavrines/Kinetic.jl#12 @rdeits
[JOSS Review] Initial installation + testing notes
Hi @vavrines--I'm working on the review for JOSS; here are a few initial notes from the installation and testing process:

* The instructions here: https://xiaotianbai.com/Kinetic.jl/dev/install/ instruct users to `add Kinetic` but also `build Kinetic` and `instantiate`. Only that first instruction is actually necessary. Adding a package automatically builds it, and `instantiate` is only necessary when installing the packages listed in a `Manifest.toml`, which is not the case here. You can simplify those instructions to `add Kinetic`

* I tried following the instructions in `paper.md`, but got an error during the call to `initialize`. The full output is below:

```
julia> set, ctr, face, t = initialize("config.toml")
==============================================================
Kinetic.jl
Portable Kinetic Simulation and Scientific Machine Learning
==============================================================

[ Info: initializing solver

[ Info: Reading config from config.toml
--------------------------------------------------------------
matter = gas # material
case = cavity # case
space = 2d2f2v # phase
flux = kfvs # flux function
collision = bgk # intermolecular collision
nSpecies = 1 # number of species
interpOrder = 2 # interpolation order of accuracy
limiter = vanleer # limiter function
boundary = maxwell # boundary condition
cfl = 0.8 # CFL number
maxTime = 5.0 # maximal simulation time
x0 = 0.0 # starting point in x
x1 = 1.0 # ending point in x
nx = 45 # number of cells in x
y0 = 0.0 # starting point in y
y1 = 1.0 # ending point in y
ny = 45 # number of cells in y
pMeshType = uniform # mesh type
nxg = 0 # number of ghost cell in x
nyg = 0 # number of ghost cell in y
umin = -5.0 # starting point in u
umax = 5.0 # ending point in u
nu = 28 # number of cells in u
vmin = -5.0 # starting point in v
vmax = 5.0 # ending point in v
nv = 28 # number of cells in v
vMeshType = rectangle # mesh type
nug = 0 # number of ghost cell in u
nvg = 0 # number of ghost cell in v
knudsen = 0.075 # Knudsen number
mach = 0.0 # Mach number
prandtl = 1.0 # Prandtl number
inK = 1.0 # molecular inner degree of freedom
omega = 0.72 # viscosity index of hard-sphere gas
alphaRef = 1.0 # viscosity index of hard-sphere gas in reference state
omegaRef = 0.5 # viscosity index of hard-sphere gas ub reference state
uLid = 0.15 # U-velocity of moving wall
vLid = 0.0 # V-velocity of moving wall
tLid = 1.0 # temperature of wall
--------------------------------------------------------------

ERROR: MethodError: no method matching Setup(::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String}, ::SubString{String})
Closest candidates are:
  Setup(::S, ::S, ::S, ::S, ::S, ::I, ::I, ::S, ::S, ::E, ::F) where {S<:AbstractString, I<:Integer, E<:Real, F<:Real} at /home/rdeits/.julia/packages/KitBase/4GfNd/src/Type/struct_general.jl:23
Stacktrace:
 [1] set_setup(dict::Dict{Symbol, Any})
   @ KitBase ~/.julia/packages/KitBase/4GfNd/src/Solver/solver_set.jl:162
 [2] SolverSet(configfilename::String)
   @ KitBase ~/.julia/packages/KitBase/4GfNd/src/Solver/solver_set.jl:54
 [3] initialize(configfilename::String)
   @ KitBase ~/.julia/packages/KitBase/4GfNd/src/Solver/solver_init.jl:32
 [4] top-level scope
   @ REPL[5]:1
```

Note that I specifically installed v0.7.0, as specified in the review: https://github.com/openjournals/joss-reviews/issues/3060
