### 2021-04-08 #3060 @rdeits
I've submitted an issue to Kinetic.jl with some notes on the installation + setup process: https://github.com/vavrines/Kinetic.jl/issues/12
