### 2021-04-08 e03e7d9
Update installation docs
