### 2021-04-08 vavrines/Kinetic.jl#12 @vavrines
Hi @rdeits

Thanks for the feedback!

- The documentation has been revised in the latest dev.
- I'm sorry for the inconvenience. The comments after `#` play as an explanation of parameters, and I just realize that they are incompatible with the I/O. A quick fix is to remove `#` and the contents afterwards. The I/O method has been rewritten here
(https://github.com/vavrines/KitBase.jl/commit/368a818768679f1f3aaf931ab730d7e61f8640ab) in the v0.4.3 of KitBase.jl.
