+++ example/mixture_shock.jl
@@ -1,4 +1,5 @@
-using ProgressMeter, OffsetArrays, Kinetic
+using Kinetic
+using KitBase.OffsetArrays, KitBase.ProgressMeter
 
 cd(@__DIR__)
 D = read_dict("mixture_shock.txt")
@@ -9,8 +10,8 @@ end
 
 begin
     γ = heat_capacity_ratio(inK, 1)
-    set = Setup(case, space, flux, collision, nSpecies, interpOrder, limiter, cfl, maxTime)
-    pSpace = PSpace1D(x0, x1, nx, pMeshType, nxg)
+    set = Setup(matter, case, space, flux, collision, nSpecies, interpOrder, limiter, boundary, cfl, maxTime)
+    pSpace = PSpace1D(x0, x1, nx, nxg)
     ue0 = umin * sqrt(mi / me)
     ue1 = umax * sqrt(mi / me)
     vSpace = MVSpace1D(umin, umax, ue0, ue1, nu, vMeshType, nug)
@@ -63,9 +64,9 @@ begin
     nt = Int(floor(ks.set.maxTime / dt))
 end
 
-Kinetic.reconstruct!(ks, ctr)
-
-Kinetic.evolve!(ks, ctr, face, dt)
-
 res = zeros(3, 2)
-Kinetic.update!(ks, ctr, face, dt, res; bc=:fix)
+begin
+    Kinetic.reconstruct!(ks, ctr)
+    Kinetic.evolve!(ks, ctr, face, dt)
+    Kinetic.update!(ks, ctr, face, dt, res; bc=:fix)
+end
