+++ example/test_briowu.jl
@@ -1,410 +1,0 @@
-using Revise, Kinetic, ProgressMeter
-
-function flux_tst!(
-    fw::AbstractArray{<:AbstractFloat,2},
-    fh0::AbstractArray{<:AbstractFloat,2},
-    fh1::AbstractArray{<:AbstractFloat,2},
-    fh2::AbstractArray{<:AbstractFloat,2},
-    fh3::AbstractArray{<:AbstractFloat,2},
-    wL::AbstractArray{<:Real,2},
-    h0L::AbstractArray{<:AbstractFloat,2},
-    h1L::AbstractArray{<:AbstractFloat,2},
-    h2L::AbstractArray{<:AbstractFloat,2},
-    h3L::AbstractArray{<:AbstractFloat,2},
-    wR::AbstractArray{<:Real,2},
-    h0R::AbstractArray{<:AbstractFloat,2},
-    h1R::AbstractArray{<:AbstractFloat,2},
-    h2R::AbstractArray{<:AbstractFloat,2},
-    h3R::AbstractArray{<:AbstractFloat,2},
-    u::AbstractArray{<:AbstractFloat,2},
-    ω::AbstractArray{<:AbstractFloat,2},
-    inK::Real,
-    γ::Real,
-    mi::Real,
-    ni::Real,
-    me::Real,
-    ne::Real,
-    Kn::Real,
-    dt::Real,
-)
-
-    #--- upwind reconstruction ---#
-    δ = heaviside.(u)
-
-    h0 = @. h0L * δ + h0R * (1.0 - δ)
-    h1 = @. h1L * δ + h1R * (1.0 - δ)
-    h2 = @. h2L * δ + h2R * (1.0 - δ)
-    h3 = @. h3L * δ + h3R * (1.0 - δ)
-
-    #w = mixture_moments_conserve(h0, h1, h2, h3, u, ω)
-    primL = mixture_conserve_prim(wL, γ)
-    primR = mixture_conserve_prim(wR, γ)
-
-    #--- construct interface distribution ---#
-    Mu1, Mv1, Mw1, MuL1, MuR1 = mixture_gauss_moments(primL, inK)
-    Muv1 = mixture_moments_conserve(MuL1, Mv1, Mw1, 0, 0, 0)
-    Mu2, Mv2, Mw2, MuL2, MuR2 = mixture_gauss_moments(primR, inK)
-    Muv2 = mixture_moments_conserve(MuR2, Mv2, Mw2, 0, 0, 0)
-
-    w = similar(wL)
-    for j in axes(w, 2)
-        @. w[:, j] = primL[1, j] * Muv1[:, j] + primR[1, j] * Muv2[:, j]
-    end
-    prim = mixture_conserve_prim(w, γ)
-
-    tau = aap_hs_collision_time(prim, mi, ni, me, ne, Kn)
-    #@. tau +=
-    #    abs(primL[1, :] / primL[end, :] - primR[1, :] / primR[end, :]) /
-    #    (primL[1, :] / primL[end, :] + primR[1, :] / primR[end, :]) *
-    #    dt *
-    #    5.0
-    mprim = aap_hs_prim(prim, tau, mi, ni, me, ne, Kn)
-
-    Mt = zeros(2, 2)
-    @. Mt[2, :] = tau * (1.0 - exp(-dt / tau)) # f0
-    @. Mt[1, :] = dt - Mt[2, :] # M0
-
-    #--- calculate fluxes ---#
-    Mu, Mv, Mw, MuL, MuR = mixture_gauss_moments(mprim, inK)
[…]
