+++ example/sod.txt
@@ -3,6 +3,7 @@
 # ------------------------------------------------------------
 
 # case
+matter = gas
 case = sod
 space = 1d2f1v
 nSpecies = 1
