### 2021-04-19 #3060 @rdeits
Thanks for the reminder. I'm still having trouble getting the basic example to run at all--I've updated the discussion here: https://github.com/vavrines/Kinetic.jl/issues/12#issuecomment-822746390
