+++ paper/paper.md
@@ -114,13 +114,13 @@ tLid = 1.0 # temperature of wall
 let us execute the following codes
 ```julia
 using Kinetic
-set, ctr, face, t = initialize("config.toml")
-t = solve!(set, ctr, face, t)
+set, ctr, xface, yface, t = initialize("config.toml")
+t = solve!(set, ctr, xface, yface, t)
 plot_contour(set, ctr)
 ```
 
 The computational setup is stored in `set`. 
-The control volume and interface solutions are stored in `ctr` and `face` correspondingly. 
+The solutions over control volumes are represented in an array `ctr`, while `xface` and `yface` record the interface fluxes along x and y directions.
 The result is visualized with built-in function `plot_contour`, which presents the distributions of gas density, velocity and temperature inside the cavity.
 
 ![Fig. 1](cavity.png)
