### 2021-04-19 vavrines/Kinetic.jl#12 @vavrines
@rdeits
I apologize for the inconvenience. I should have checked it again.
The example codes should be as follows
```julia
using Kinetic
set, ctr, xface, yface, t = initialize("config.toml")
t = solve!(set, ctr, xface, yface, t)
plot_contour(set, ctr)
```
where `xface` and `yface` record numerical fluxes along x and y directions respectively.
I've revised the paper, and the tutorials in the documentation (https://xiaotianbai.com/Kinetic.jl/dev/tutorial/) should all work fine.
