### 2021-04-19 #3060 @vavrines
@diehlpk
Dear Patrick, May I have a question here?
I found the JOSS doesn't recognize my markdown bullet points at the end of first page, i.e.
```
- 100% Julia stack that encounters no two-language problem
- Comprehensive support for kinetic theory and phase-space equations
- ...
```
Does JOSS support any format of bullet points, or I have to reformat it into plain text?
Best,
