### 2021-04-20 #3060 @vavrines
> @vavrines We used in our JOSS paper both the dash and star. Please have a look at the Application section
>
> https://github.com/STEllAR-GROUP/hpx/blob/master/docs/joss_paper/paper.md
>
> Can you try to have a blank line and indent the dash or star?

With blank line it works, thx a lot!
