### 2021-05-04 #2942 @diehlpk
Hey @ChrisRackauckas and @dpsanders would you be interested to review this paper?
