### 2021-05-06 #2942 @jarvist
> Hey @Adrianzo @ahwillia @vchuravy @dhhagan @jarvist would you be interested to review this paper?

I'm on parental leave until May 17th, but can then review this.
We have some partial reviews already? The history ^^^^ was a bit long to skim through!
