### 2021-05-11 3d5f2bf
CompatHelper: bump compat for "KitBase" to "0.5"
