+++ Project.toml
@@ -9,7 +9,7 @@ KitML = "b2531deb-9288-45c6-8f4a-0220551da45d"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
-KitBase = "0.4"
+KitBase = "0.4, 0.5"
 KitML = "0.3"
 Reexport = "0.2, 1.0"
 julia = "^1"
