### 2021-05-12 #2942 @sloede
I have never reviewed at JOSS, how does it compare to a regular paper review in terms of effort? Also, what is the expected time frime? I am intrigued, but quite busy this and next week...
