### 2021-05-12 #2942 @vavrines
@sloede Thanks for the interest : )
It's depends on a number of factors and hard to make a fair comparison. From my point of view on the checklist given by JOSS, it's easier to review here.
cc @diehlpk I remember the expected time frime you mentioned is 2 months. Given the long delay of the review process, I'm not sure if it stays the same in this round.
