### 2021-05-18 #3060 @rdeits
I was able to get all the examples running properly, so I should be able to
finish this weekend.

On Tue, May 18, 2021 at 2:36 PM Patrick Diehl ***@***.***>
wrote:

> Hey @rdeits <https://github.com/rdeits> when do you estimate to finish
> the review?
>
