### 2021-05-21 b8503fd
CompatHelper: bump compat for "KitBase" to "0.5"
