### 2021-05-23 #3060 @rdeits
Could one of the `openjournals` admins confirm that I still have collaborator access on this repository? I can no longer modify the checklist above, and I do not see this repo listed under my https://github.com/settings/repositories list anymore.

I know that I previously had access because I was able to partially complete the checklist earlier.
