### 2021-06-01 #3060 @diehlpk
@rdeits Please click on the link above and you should be able to edit the checkboxes again.
