### 2021-06-05 #3060 @rdeits
@diehlpk yes, thanks--the invite worked and I'm now able to edit the checklist again. I'm not sure how I lost access to begin with, but I'm glad it's all sorted now.
