### 2021-06-05 vavrines/Kinetic.jl#12 @rdeits
Thanks again for updating the examples--all of the code now appears to work exactly as expected.

I just have two requests before finishing the review checklist:

1. Could you please add the exact example you posted above to the `Kinetic.jl/test/runtests.jl`? (currently the `initialize` call is covered, but not `solve!` or `plot_contour`. I see that KitBase.jl and the other modules already have good test coverage, but having a complete example in the Kinetic.jl tests should reduce the likelihood of that example accidentally breaking in the future again.
2. There are a few minor errors in the paper--could you please take a look?
  * Could you add one sentence to the summary about what the role of machine learning is here? For example, I could imagine using optimization or ML to try to design a cavity to achieve some fluid property. Or I could imagine using ML to try to approximate the entire fluid dynamics calculation to improve performance. Or perhaps something completely different.
  * "KitBase.jl with basic physics" should be "KitBase.jl for basic physics", and likewise "KitML.jl for neural dynamics"
  * I'm not sure what the intended meaning of "hooked" is in "neural equation can be hooked and solved". I would suggest "can be set up and solved" instead.
  * "algorithmic efficiency for application" should be "algorithmic efficiency for applications"
  * "are splitted" should be "are split" and "is splitted" should be "is split". I have no idea why the past tense of "split" is "split"--English is just kind of weird that way.
  * "for broad devices" should probably be "for a broad range of devices"
  * "Kinetic.jl focus" should be "Kinetic.jl focuses"
  * "holds the perfect possibility" is a bit awkward. "is possible" might be better there.
  * "It depends Flux.jl" should be "it uses Flux.jl" or "it depends on Flux.jl"
  * "Closely coupling" should probably be "Close coupling" or "Easy integration"
  * "we've develop" should be "we've developed" (or probably "we have developed" since contractions like "we've" are less common in formal writing like this)
  * "from Python ecosystem" should be "from the Python ecosystem"
