### 2021-06-05 #3060 @rdeits
Ok, I think this is in good shape. I made a couple of final requests here: https://github.com/vavrines/Kinetic.jl/issues/12#issuecomment-855258289 (mostly just minor typos for grammatical issues), and I've checked off everything else in the list.
