### 2021-06-06 1e127f4
Add a complete example in test
