+++ test/runtests.jl
@@ -1,3 +1,10 @@
-using Test, Kinetic
+using Kinetic
 
-ks, ctr, face, simTime = Kinetic.initialize("config.txt")
+# initialization
+set, ctr, xface, yface, t = initialize("config.toml")
+
+# solution algorithm
+t = solve!(set, ctr, xface, yface, t)
+
+# visualization
+plot_contour(set, ctr)
