### 2021-06-06 c2068ba
Update paper.md
Fix syntax errors
