+++ paper/paper.bib
@@ -57,7 +57,7 @@
   biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1811-01457},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
-@misc{xiao2020,
+@misc{xiao2020b,
   title={Using neural networks to accelerate the solution of the Boltzmann equation}, 
   author={Tianbai Xiao and Martin Frank},
   year={2020},
@@ -72,5 +72,14 @@
   volume={332},
   pages={475--491},
   year={2017},
+  publisher={Elsevier}
+}
+@article{xiao2020a,
+  title={A velocity-space adaptive unified gas kinetic scheme for continuum and rarefied flows},
+  author={Xiao, Tianbai and Liu, Chang and Xu, Kun and Cai, Qingdong},
+  journal={Journal of Computational Physics},
+  volume={415},
+  pages={109535},
+  year={2020},
   publisher={Elsevier}
 }
