+++ paper/paper.bib
@@ -57,7 +57,7 @@
   biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1811-01457},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
-@misc{xiao2020b,
+@misc{xiao2020,
   title={Using neural networks to accelerate the solution of the Boltzmann equation}, 
   author={Tianbai Xiao and Martin Frank},
   year={2020},
