### 2021-06-07 #3060 @diehlpk
> Hi @diehlpk
>
> I have a problem compiling my paper, which is related to the LaTex package `{hyperref}`.
> I didn't cite a `jarvist` in my markdown file nor in the bibtex.
> It seems that whedon somehow misunderstood the reviewer as a citation.
> Do you have an idea on how to fix it?

@vavrines I asked the support team for help, since I do not know what is going wrong.
