### 2021-06-07 #3060 @arfon
> Submitting author: @vavrines (Tianbai Xiao)
> Repository: https://github.com/vavrines/Kinetic.jl
> Version: v0.7.0
> Editor: @diehlpk
> Reviewers: @rdeits @jarvist, @rdeits

The issue here is that the `Reviewers:` string `@rdeits @jarvist, @rdeits` is missing a comma, and so Whedon is passing `@jarvist` as a variable into the compilation step which is being interpreted as a citation.

So, the fix here is to add a comma between `@rdeits @jarvist` in the reviewer list above (which I'm going to do now).
