### 2021-06-07 vavrines/Kinetic.jl#12 @vavrines
Hi @rdeits

Thank you very much for the input!
The manuscript has been revised with the incorporation of all these suggestions, including

- The complete workflow is added into `runtests.jl`.
- A one-sentence summary of machine learning is added into the summary.
- The grammatical errors are fixed.

Please find the revised paper here (https://github.com/openjournals/joss-papers/blob/joss.03060/joss.03060/10.21105.joss.03060.pdf).
