### 2021-06-13 vavrines/Kinetic.jl#12 @rdeits
Thanks for the updates--it all looks good to me!
