### 2021-06-14 #3060 @diehlpk
@rdeits Thanks, I just checked, and it seems that `Quality of writing` is unchecked. Could you please have a look?
