### 2021-06-14 #3060 @jarvist
I seem to have lost my 'collaborator' badge, and so cannot tick off things on my checklist.
I'm just reviewing this package and paper now, so will make a note separately as I go.
