### 2021-06-14 #3060 @jarvist
@vavrines - it's a bit suboptimal that Kinetic (rather KitBase) has a dependency on Plots. Currently this has actually broken precompilation on my machine, due to some kind of interaction with the way Plots is being initialised & attempting to pull in GR (which is a binary package, and breaks not irregularly).

It would require quite a lot of refactoring, but using PlotsRecipes avoids this issue, and makes the dependency lighter to install.
https://docs.juliaplots.org/latest/recipes/
