### 2021-06-14 #3060 @jarvist
@vavrines Could you please confirm that the example given in the JOSS paper works OK?
I'm currently getting:
```
; julia
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _` |  |
  | | |_| | | | (_| |  |  Version 1.6.1 (2021-04-23)
 _/ |\__'_|_|_|\__'_|  |  Official https://julialang.org/ release
|__/                   |

julia> using Kinetic

julia> set, ctr, xface, yface, t = Kinetic.initialize("kinetic.toml")
==============================================================
Kinetic.jl
A Lightweight Toolbox for Kinetic Modeling and Simulation
==============================================================

reading configurations from kinetic.toml

initializing solver:
ERROR: MethodError: no method matching iterate(::Nothing)
Closest candidates are:
  iterate(::Union{LinRange, StepRangeLen}) at range.jl:664
  iterate(::Union{LinRange, StepRangeLen}, ::Int64) at range.jl:664
  iterate(::T) where T<:Union{Base.KeySet{var"#s79", var"#s78"} where {var"#s79", var"#s78"<:Dict}, Base.ValueIterator{var"#s77"} where var"#s77"<:Dict} at dict.jl:693
  ...
Stacktrace:
 [1] indexed_iterate(I::Nothing, i::Int64)
   @ Base ./tuple.jl:89
 [2] top-level scope
   @ REPL[2]:1

julia>
```
