### 2021-06-14 #3060 @vavrines
@jarvist
I Just tested the code again. Actually I never saw such error beforehand. It seems that Julia doesn't recognize any text in the config file (if it's not empty).
I'm not sure how you generate the config (maybe an encoding thing?), but I guess a copy-paste of the CI config file will help get rid of this.
Please let me know if it doesn't work.
