### 2021-06-14 #3060 @jarvist
@vavrines Thanks! It seems I ended up pinned on an old version of Kinetic due to version incompatibility with other installed packages. I've got that sorted now, and the example from the JOSS (copied + pasted out of the PDF, as I assumed that someone would try that at some point...) seems to be solving.
