+++ versions.js
@@ -5,3 +5,5 @@ var DOC_VERSIONS = [
   "v0.5",
   "dev",
 ];
+var DOCUMENTER_NEWEST = "v0.7.4";
+var DOCUMENTER_STABLE = "stable";
