### 2021-06-14 #3060 @jarvist
OK! All looking good.
Two minor issues:-
 * the 'community guidelines' are OK, but a little unclear. How should people get in touch? via a Github issue? Personal email? etc.
 * the aforementioned dependency on Plots.jl; this would be better refactored to use Plots Recipes

But I don't think either of these things should block publication.
