### 2021-06-14 #3060 @vavrines
> OK! All looking good.
> Two minor issues:-
>
> * the 'community guidelines' are OK, but a little unclear. How should people get in touch? via a Github issue? Personal email? etc.
> * the aforementioned dependency on Plots.jl; this would be better refactored to use Plots Recipes
>
> But I don't think either of these things should block publication.

Thanks for the input! I'm gonna make the point clearer.
The RecipesBase idea sounds interesting. I'm not sure whether it's the best fit here, since the `plot_line` and `plot_contour` functions need to use a plot function anyway. I'll look into it.
