### 2021-06-14 #3060 @diehlpk
@rdeits and @jarvist thanks for reviewing for JOSS!
