### 2021-06-14 #3060 @diehlpk
@vavrines Please address my editorial issues below and we are good to prepare the final paper.

1. Karlsruhe Institute of Technology -> Please add the city and country to the affiliation
2.  `For example, it is cumbersome to implement the phase-23field evolution from the Boltzmann equation (Xiao et al., 2017,2020) in OpenFOAM or24integrate it with scientific machine learning (SciML) package` Please elaborate here and give some explanation why it is cumbersome.
3.  `Besides, the two-language problem introduces additional trade-off in both development and execution` Please elaborate on the trade off.
4. I think some reference to Julia itself would be beneficial for the reader. maybe> https://github.com/JuliaLang/julia/blob/master/CITATION.bib
