+++ paper/paper.bib
@@ -3,12 +3,14 @@
   author={Chapman, Sydney and Cowling, Thomas George and Burnett, David},
   year={1990},
   publisher={Cambridge university press}
+  doi={10.2307/3609795}
 }
 @book{batchelor2000,
   title={An introduction to fluid dynamics},
   author={Batchelor, Cx K and Batchelor, GK},
   year={2000},
   publisher={Cambridge university press}
+  doi={10.1017/CBO9780511800955}
 }
 @article{bezanson2017,
   title={Julia: A fresh approach to numerical computing},
@@ -19,6 +21,7 @@
   pages={65--98},
   year={2017},
   publisher={SIAM}
+  doi={10.1137/141000671}
 }
 @inproceedings{jasak2007,
   title={OpenFOAM: A C++ library for complex physics simulations},
@@ -34,7 +37,7 @@
   author={{Clawpack Development Team}},
   url={http://www.clawpack.org},
   note={Version 5.7.1},
-  doi={https://doi.org/10.5281/zenodo.4025432},
+  doi={10.5281/zenodo.4025432},
   year={2020}}
 @article{Flux2018,
   author    = {Michael Innes and
@@ -73,6 +76,7 @@
   pages={475--491},
   year={2017},
   publisher={Elsevier}
+  doi={10.1016/j.jcp.2016.12.022}
 }
 @article{xiao2020a,
   title={A velocity-space adaptive unified gas kinetic scheme for continuum and rarefied flows},
@@ -82,4 +86,5 @@
   pages={109535},
   year={2020},
   publisher={Elsevier}
+  doi={10.1016/j.jcp.2020.109535}
 }
