+++ paper/paper.bib
@@ -12,6 +12,16 @@
   publisher={Cambridge university press}
   doi={10.1017/CBO9780511800955}
 }
+@article{zhu2017dugksfoam,
+  title={dugksFoam: An open source OpenFOAM solver for the Boltzmann model equation},
+  author={Zhu, Lianhua and Chen, Songze and Guo, Zhaoli},
+  journal={Computer Physics Communications},
+  volume={213},
+  pages={155--164},
+  year={2017},
+  publisher={Elsevier}
+  doi={10.1016/j.cpc.2016.11.010}
+}
 @article{bezanson2017,
   title={Julia: A fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
