+++ paper/paper.bib
@@ -12,7 +12,7 @@
   publisher={Cambridge university press}
   doi={10.1017/CBO9780511800955}
 }
-@article{zhu2017dugksfoam,
+@article{zhu2017,
   title={dugksFoam: An open source OpenFOAM solver for the Boltzmann model equation},
   author={Zhu, Lianhua and Chen, Songze and Guo, Zhaoli},
   journal={Computer Physics Communications},
