+++ paper/paper.bib
@@ -98,3 +98,13 @@
   publisher={Elsevier}
   doi={10.1016/j.jcp.2020.109535}
 }
+@article{krause2021,
+  title={OpenLB—Open source lattice Boltzmann code},
+  author={Krause, Mathias J and Kummerl{\"a}nder, Adrian and Avis, Samuel J and Kusumaatmaja, Halim and Dapelo, Davide and Klemens, Fabian and Gaedtke, Maximilian and Hafen, Nicolas and Mink, Albert and Trunk, Robin and others},
+  journal={Computers \& Mathematics with Applications},
+  volume={81},
+  pages={258--288},
+  year={2021},
+  publisher={Elsevier}
+  doi={10.1016/j.camwa.2020.04.033}
+}
