+++ paper/paper.bib
@@ -2,14 +2,14 @@
   title={The mathematical theory of non-uniform gases: an account of the kinetic theory of viscosity, thermal conduction and diffusion in gases},
   author={Chapman, Sydney and Cowling, Thomas George and Burnett, David},
   year={1990},
-  publisher={Cambridge university press}
+  publisher={Cambridge university press},
   doi={10.2307/3609795}
 }
 @book{batchelor2000,
   title={An introduction to fluid dynamics},
   author={Batchelor, Cx K and Batchelor, GK},
   year={2000},
-  publisher={Cambridge university press}
+  publisher={Cambridge university press},
   doi={10.1017/CBO9780511800955}
 }
 @article{zhu2017,
@@ -19,7 +19,7 @@
   volume={213},
   pages={155--164},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.cpc.2016.11.010}
 }
 @article{bezanson2017,
@@ -30,7 +30,7 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
   doi={10.1137/141000671}
 }
 @inproceedings{jasak2007,
@@ -85,7 +85,7 @@
   volume={332},
   pages={475--491},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.jcp.2016.12.022}
 }
 @article{xiao2020a,
@@ -95,7 +95,7 @@
   volume={415},
   pages={109535},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.jcp.2020.109535}
 }
 @article{krause2021,
@@ -105,7 +105,7 @@
   volume={81},
   pages={258--288},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.camwa.2020.04.033}
 }
 @article{bezanson2012julia,
