+++ paper/paper.bib
@@ -108,7 +108,7 @@
   publisher={Elsevier},
   doi={10.1016/j.camwa.2020.04.033}
 }
-@article{bezanson2012julia,
+@article{bezanson2012,
   title={Julia: A fast dynamic language for technical computing},
   author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
   journal={arXiv preprint arXiv:1209.5145},
