### 2021-06-15 #3060 @vavrines
Hi @diehlpk

Thank you very much for the suggestions.
I revised the paper with the incorporation of the editorial issues.
The major changes include:

- The city and country were added to the affiliation.
- The two points were reorganized and elaborated to avoid misunderstanding.
- A Julia paper was already cited in the first version, and I added the original paper in 2012 in the bibliography.

Please have a look at the revision and let me know if there are further ambiguities.
