### 2021-06-15 #3060 @diehlpk
@vavrines I found some typos in the latest version of the paper

* cubersome -> cumbersome
* optmizing - > optimizing
* vectorizaion -> vectorization

Please fix them before we can go forward.
