### 2021-06-15 d875175
Fix typos in paper
