### 2021-06-15 #3060 @diehlpk
@vavrines

- [ ] Generate a new release with all fixed in the paper and the code made during the review. Please post the release number here.
- [ ] Upload the release to Zendo or Figshare and post the DOI here.
