### 2021-06-15 #3060 @vavrines
@diehlpk

- The new release number is v0.7.5 (https://github.com/vavrines/Kinetic.jl/releases/tag/v0.7.5), with the title same as paper.
- The Zenodo record is here (https://zenodo.org/record/4958132#.YMjH9C0RppR)
