### 2021-06-15 #3060 @diehlpk
@vavrines

`vavrines/Kinetic.jl: Kinetic.jl: A portable finite volume toolbox for scientific and neural computing`

The Zendo title does not match the title of the paper. Please update the title there as well. I recommend to add your DOI asl well.
