+++ paper/paper.md
@@ -19,42 +19,42 @@ bibliography: paper.bib
 # Summary
 
 Kinetic.jl is a lightweight finite volume toolbox written in the Julia programming language for the study of computational physics and scientific machine learning.
-It is an open-source project hosted on GitHub and distributed under MIT license.
+It is an open-source project hosted on GitHub and distributed under the MIT license.
 The main module consists of KitBase.jl for basic physics and KitML.jl for neural dynamics.
 The function library provides a rich set of numerical fluxes and source terms for differential and integral equations.
 Any advection-diffusion type mechanical or neural equation can be set up and solved within the framework.
-The machine learning methods can be integrated seamlessly to build data-driven closure models and accelerate the calculation of nonlinear terms.
-The package is designed to balance the programming flexibility for scientific research, the algorithmic efficiency for applications, and the simplicity for educational usage.
+Machine learning methods can be seamlessly integrated to build data-driven closure models and accelerate the calculation of nonlinear terms.
+The package is designed to balance programming flexibility for scientific research, algorithmic efficiency for applications, the simplicity for educational usage.
 
 # Statement of need
 
 A physical system can perform a wonderfully diverse set of acts on different characteristic scales.
-It is challenging to propose a universal theory that can be applied for describing multi-scale physical evolutions quantitatively.
-For example, the particle transports can be depicted statistically by fluid mechanics at a macroscopic level [@batchelor2000], but needs to be followed in more detail by the kinetic theory of gases at molecular mean free path scale [@chapman1990].
-With the rapidly advancing computing power, the finite volume method (FVM) provides a prevalent method to conduct direct numerical simulations based on first physical principles.
+It is challenging to propose a universal theory that can be applied to describing multi-scale physical evolutions quantitatively.
+For example, particle transport can be depicted statistically by fluid mechanics at a macroscopic level [@batchelor2000], but needs to be followed in more detail by the kinetic theory of gases at the molecular mean free path scale [@chapman1990].
+With rapidly advancing computing power, the finite volume method (FVM) provides a prevalent method to conduct direct numerical simulations based on first physical principles.
 
-Most of the existing FVM libraries, e.g. the OpenFOAM [@jasak2007], are dedicated to solving the Euler and the Navier-Stokes equations.
+Most existing FVM libraries, e.g., OpenFOAM [@jasak2007], are dedicated to solving the Euler and the Navier-Stokes equations.
 Very limited work has been done for phase-field models [@zhu2017; @krause2021].
 Since classical fluid dynamics basically requires an one-shot simulation process from initial to final solution fields, these libraries are mostly written in compiled languages (C/C++ and Fortran).
-Such approach enjoy the perfect execution efficiency but sacrifices more or less the flexibility of secondary development.
-It becomes cumbersome to integrate the existing numerical solver with scientific machine learning (SciML) packages, as interactive programming is becoming a mainstream practice in data science.
-It also brings considerable difficulties to general or educational users who are not familiar with the package in configuring environments and compiling binaries.
+Such approaches enjoy good execution efficiency but sacrifice the flexibility of secondary development.
+This makes it cumbersome to integrate existing numerical solvers with scientific machine learning (SciML) packages, as interactive programming is becoming a mainstream practice in data science.
+This also causes considerable difficulties to general or educational users who are not familiar with the package in configuring environments and compiling binaries.
 
-One compromise can be made with a combination of static and dynamic languages [@clawpack2020], where the high-level front-ends and the low-level computational back-ends are split.
[…]
