+++ paper/paper.bib
@@ -2,18 +2,18 @@
   title={The mathematical theory of non-uniform gases: an account of the kinetic theory of viscosity, thermal conduction and diffusion in gases},
   author={Chapman, Sydney and Cowling, Thomas George and Burnett, David},
   year={1990},
-  publisher={Cambridge university press},
+  publisher={Cambridge University Press},
   doi={10.2307/3609795}
 }
 @book{batchelor2000,
   title={An introduction to fluid dynamics},
-  author={Batchelor, Cx K and Batchelor, GK},
+  author={Batchelor, George K.},
   year={2000},
-  publisher={Cambridge university press},
+  publisher={Cambridge University Press},
   doi={10.1017/CBO9780511800955}
 }
 @article{zhu2017,
-  title={dugksFoam: An open source OpenFOAM solver for the Boltzmann model equation},
+  title={dugksFoam: An open source {OpenFOAM} solver for the {B}oltzmann model equation},
   author={Zhu, Lianhua and Chen, Songze and Guo, Zhaoli},
   journal={Computer Physics Communications},
   volume={213},
@@ -34,7 +34,7 @@
   doi={10.1137/141000671}
 }
 @inproceedings{jasak2007,
-  title={OpenFOAM: A C++ library for complex physics simulations},
+  title={OpenFOAM: A {C}++ library for complex physics simulations},
   author={Jasak, Hrvoje and Jemcov, Aleksandar and Tukovic, Zeljko and others},
   booktitle={International workshop on coupled methods in numerical dynamics},
   volume={1000},
@@ -59,7 +59,7 @@
                Tejan Karmali and
                Avik Pal and
                Viral Shah},
-  title     = {Fashionable Modelling with Flux},
+  title     = {Fashionable Modelling with {F}lux},
   journal   = {CoRR},
   volume    = {abs/1811.01457},
   year      = {2018},
@@ -71,7 +71,7 @@
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 @misc{xiao2020b,
-  title={Using neural networks to accelerate the solution of the Boltzmann equation}, 
+  title={Using neural networks to accelerate the solution of the {B}oltzmann equation}, 
   author={Tianbai Xiao and Martin Frank},
   year={2020},
   eprint={2010.13649},
@@ -99,7 +99,7 @@
   doi={10.1016/j.jcp.2020.109535}
 }
 @article{krause2021,
-  title={OpenLB—Open source lattice Boltzmann code},
+  title={OpenLB—Open source lattice {B}oltzmann code},
   author={Krause, Mathias J and Kummerl{\"a}nder, Adrian and Avis, Samuel J and Kusumaatmaja, Halim and Dapelo, Davide and Klemens, Fabian and Gaedtke, Maximilian and Hafen, Nicolas and Mink, Albert and Trunk, Robin and others},
   journal={Computers \& Mathematics with Applications},
   volume={81},
