### 2021-06-15 #3060 @danielskatz
@vavrines - While proofreading this paper before publishing it, I found a number of things that I suggested fixes for in https://github.com/vavrines/Kinetic.jl/pull/15  Please merge this or let me know what you disagree with, then we can proceed with the final publication.
