### 2021-06-15 #3060 @vavrines
> @vavrines - While proofreading this paper before publishing it, I found a number of things that I suggested fixes for in [vavrines/Kinetic.jl#15](https://github.com/vavrines/Kinetic.jl/pull/15) Please merge this or let me know what you disagree with, then we can proceed with the final publication.

Hi @danielskatz

Thank you very much for the careful check and revision.
The patch has been merged into master branch.
Do I need to make another release on Zenodo for the latest change?
