### 2021-06-15 #3060 @danielskatz
no, you don't need to make a new release for changes in the paper/bib
